backend/requirements/base.txt
psycopg2-binary==2.9.6
environs[django]==9.5.0
requests==2.28.2 
markdown==3.4.3       # Markdown support for the browsable API.
# Django
# ------------------------------------------------------------------------------
django==4.2
djangorestframework==3.14.0
django-filter==23.1
dj-rest-auth==4.0.1
django-allauth==0.54.0